backend/requirements.txt
# FastAPI 웹 프레임워크 (백엔드 API 구축용)
fastapi==0.104.1

# ASGI 서버, FastAPI 앱 실행용 (standard 옵션: uvloop, httptools 포함)
uvicorn[standard]==0.24.0

# 비동기 HTTP 클라이언트, 외부 API 호출용
httpx==0.25.1

# 데이터 처리 라이브러리, 표/테이블 처리
pandas==2.1.3

# 수치 계산 라이브러리, 행렬/벡터 연산
numpy==1.26.2

# 환경변수 관리 (예: .env 파일 읽기)
python-dotenv==1.0.0

# 데이터 모델/검증 라이브러리
pydantic==2.5.0

# Pydantic 기반 설정 관리
pydantic-settings==2.1.0

# 스케줄러, 정기 작업 수행
apscheduler==3.10.4

# Anthropic Claude API 클라이언트 (AI 분석)
anthropic==0.7.1

google-generativeai==0.3.2

# 버전 충돌로 패키지 설치가 불가한 경우 :
# >>> pip install --upgrade -r requirements.txt